rejiggered Schema again; create geometry works
test for creating linked geometry, and creating failed geometry passes.

Also rejiggered the Schema again, was flailing about trying to get it
to work etc etc, then realized it was failing because the test data
didn't conform to the new hpms.hpms... schema...that is, I moved the
HPMS tables into their own schema (called hpms) in the database, but I
didn't update the data I am using in the test to create the test
database.

So, also updated the test data in "hpms_test_db.sql" and the table
definitions in create_hpms_table.sql