add matched_name_from, matched_name_to
matching the road name at both the from intersection and the to
intersection, which means that maybe they aren't the same, etc etc.

It is possible that I get some sort of road, but that it is to an
incorrect destination or origin.

More likely, it is possible to get a non-road, (no geometry) because
there is no good match at the origin intersection and/or at the
destination intersection